Add support for custom validation of lines.
Due to the transformation, there can be cases where JSHint finds an issue which
is not true for the original line. One such example is the maxlen rule, where
the line containing JSX can be within the given length but the transformed JS
is too long.

When an error is found, it now checks to see if there is a custom validation
function for the given error code. The function takes the original line and the
transformed line as parameters and should return whether or not the linter can
consider the line to be valid.

A validation function for the mentioned maxlen rule has been included.